Git plugin: fix crash on commit error
When committing a small amount of files, the RunCommit() method was not returning the correct boolean status.

In case of an error on commit (for instance, user.name/user.email not set) the code was still trying to parse and display the output from the 'git commit' command, which would crash!

(in 4.30: PR #4650: Contributed by SRombauts)