Stop auto-pulling when git push fails, ask user to resolve externally
I found that in practice too often there was an "unlinking" error which was down to
git lfs not being able to replace files UE had open, which left the repo mid-rebase
which is nasty for people not familiar how to resolve. Best to be safe & avoid.
However this now displays a dialog to assist with the process.