Smarter Sync/Pull
Instead of trying to unload all the packages and reload, since this can crash Unreal in some cases, a more sensible is presented on this commit.

First we do a fetch, afterwards we read the changed files on index, and also check which files we have unstaged.
Grab the packages for those files and unload the packages.
Then the normal workflow carries on, doing the pull --rebase, and reloading the unloaded packages.

The only missing point right now is not unloading files on any local commit not yet on remote, but this should a lesser evil then unloading all the files.